combined kick_ball and local_multiplayer examples into one
