+++ examples/local_multiplayer.rs
@@ -8,75 +8,88 @@ use bevy::{
     input::gamepad::{GamepadConnection, GamepadConnectionEvent},
     prelude::*,
 };
-use bevy_enhanced_input::prelude::*;
+use bevy_enhanced_input::prelude::{Press, *};
+
+const BORDER_WIDTH: f32 = 650.0;
+const STROKE_WIDTH: f32 = 5.0;
+const BALL_RAD: f32 = 16.0;
+const ACCELERATION: f32 = 150.0;
+const KICK_IMPULSE: f32 = 1000.0;
+const FRICTION: f32 = 25.0;
+const KINETIC_FRICTION_COEFF: f32 = 3.0;
+const BOUNCINESS: f32 = 0.8;
 
 fn main() {
     App::new()
         .add_plugins((DefaultPlugins, EnhancedInputPlugin))
         .add_input_context::<Player>()
-        .add_observer(apply_movement)
-        .add_systems(Startup, spawn)
-        .add_systems(Update, update_gamepads)
+        .add_systems(Startup, setup)
+        .add_systems(Update, (calculate_physics, update_gamepads))
+        .add_observer(apply_roll)
+        .add_observer(apply_kick)
         .run();
 }
 
-fn spawn(
+fn setup(
     mut commands: Commands,
     gamepads: Query<Entity, With<Gamepad>>,
     mut meshes: ResMut<Assets<Mesh>>,
-    mut materials: ResMut<Assets<StandardMaterial>>,
+    mut materials: ResMut<Assets<ColorMaterial>>,
 ) {
+    commands.spawn(Camera2d);
+
+    // Border
+    let border_mesh = meshes.add(Rectangle::new(BORDER_WIDTH, STROKE_WIDTH));
+    let border_mat = materials.add(Color::WHITE);
+    // Bottom
     commands.spawn((
-        Camera3d::default(),
-        Transform::from_xyz(0.0, 25.0, 0.0).looking_at(-Vec3::Y, Vec3::Y),
+        Mesh2d(border_mesh.clone()),
+        MeshMaterial2d(border_mat.clone()),
+        Transform::from_translation(Vec3::Y * (-BORDER_WIDTH / 2.0)),
+    ));
+    // Top
+    commands.spawn((
+        Mesh2d(border_mesh.clone()),
+        MeshMaterial2d(border_mat.clone()),
+        Transform::from_translation(Vec3::Y * (BORDER_WIDTH / 2.0)),
+    ));
+    // Left
+    commands.spawn((
+        Mesh2d(border_mesh.clone()),
+        MeshMaterial2d(border_mat.clone()),
+        Transform::from_translation(Vec3::X * (-BORDER_WIDTH / 2.0))
+            .with_rotation(Quat::from_rotation_z(90.0f32.to_radians())),
+    ));
+    // Right
+    commands.spawn((
+        Mesh2d(border_mesh),
+        MeshMaterial2d(border_mat),
+        Transform::from_translation(Vec3::X * (BORDER_WIDTH / 2.0))
+            .with_rotation(Quat::from_rotation_z(90.0f32.to_radians())),
     ));
 
-    commands.spawn((
-        Mesh3d(meshes.add(Plane3d::new(Vec3::Y, Vec2::splat(10.0)))),
-        MeshMaterial3d(materials.add(Color::WHITE)),
-    ));
-    commands.spawn((
-        PointLight {
-            shadows_enabled: true,
-            ..Default::default()
-        },
-        Transform::from_xyz(4.0, 8.0, 4.0),
-    ));
-
-    // By default actions read inputs from all gamepads,
-    // but for local multiplayer we need assign specific
-    // gamepad index.
     let mut gamepads = gamepads.iter();
     let (gamepad1, gamepad2) = (gamepads.next(), gamepads.next());
-    let capsule = meshes.add(Capsule3d::new(0.5, 2.0));
+    let ball_mesh = meshes.add(Circle::new(BALL_RAD));
 
-    // Spawn two players with different controls.
+    // Player 1
+    let p1_mat = materials.add(Color::srgb(0.1, 0.1, 0.9));
     commands.spawn(player_bundle(
         Player::First,
         gamepad1,
-        capsule.clone(),
-        materials.add(Color::srgb_u8(124, 144, 255)),
-        Transform::from_xyz(0.0, 1.5, 8.0),
+        ball_mesh.clone(),
+        p1_mat,
+        Transform::from_xyz(-80.0, 0.0, 0.0),
     ));
+
+    let p2_mat = materials.add(Color::srgb(0.9, 0.1, 0.1));
     commands.spawn(player_bundle(
         Player::Second,
         gamepad2,
-        capsule,
-        materials.add(Color::srgb_u8(220, 90, 90)),
-        Transform::from_xyz(0.0, 1.5, -8.0),
+        ball_mesh,
+        p2_mat,
+        Transform::from_xyz(80.0, 0.0, 0.0),
     ));
-}
-
-fn apply_movement(movement: On<Fire<Movement>>, mut players: Query<&mut Transform>) {
-    let mut transform = players.get_mut(movement.context).unwrap();
-
-    // Adjust axes for top-down movement.
-    transform.translation.z -= movement.value.x;
-    transform.translation.x -= movement.value.y;
-
-    // Prevent from moving out of plane.
-    transform.translation.z = transform.translation.z.clamp(-10.0, 10.0);
-    transform.translation.x = transform.translation.x.clamp(-10.0, 10.0);
 }
 
 fn update_gamepads(
@@ -112,41 +125,129 @@ fn update_gamepads(
 fn player_bundle(
     player: Player,
     gamepad: Option<Entity>,
-    mesh: impl Into<Mesh3d>,
-    material: impl Into<MeshMaterial3d<StandardMaterial>>,
+    mesh: impl Into<Mesh2d>,
+    material: impl Into<MeshMaterial2d<ColorMaterial>>,
     transform: Transform,
 ) -> impl Bundle {
-    // Assign different bindings based on the player index.
     let move_bindings = match player {
         Player::First => Bindings::spawn((Cardinal::wasd_keys(), Axial::left_stick())),
         Player::Second => Bindings::spawn((Cardinal::arrows(), Axial::left_stick())),
+    };
+    // Duplicate manually because `SpawnRelatedBundle` doesn't implement clone
+    let kick_bindings = match player {
+        Player::First => Bindings::spawn((Cardinal::wasd_keys(), Axial::left_stick())),
+        Player::Second => Bindings::spawn((Cardinal::arrows(), Axial::left_stick())),
+    };
+
+    let arm_kick_binding = match player {
+        Player::First => bindings![KeyCode::Space],
+        Player::Second => bindings![KeyCode::ShiftRight],
     };
 
     (
         player,
         GamepadDevice::from(gamepad),
+        PlayerPhysics::default(),
         mesh.into(),
         material.into(),
         transform,
-        actions!(
-            Player[(
-                Action::<Movement>::new(),
-                DeadZone::default(),
-                SmoothNudge::default(),
+        Actions::<Player>::spawn(SpawnWith(|context: &mut ActionSpawner<_>| {
+            context.spawn((
+                Action::<Roll>::new(),
                 DeltaScale::default(),
-                Scale::splat(10.0),
+                Scale::splat(ACCELERATION),
                 move_bindings,
-            )]
-        ),
+            ));
+
+            // For controller: Kick by flicking the controller stick
+            context.spawn((
+                Action::<Kick>::new(),
+                // Replicate feel of Smash Bros., which requires the controller
+                // to flick in ~2 frames (1/30 second) for certain actions
+                Flick::new(0.0333).with_actuation(0.9),
+                Bindings::spawn(Axial::left_stick()),
+            ));
+
+            // As an alternative for keyboard players, kick by "arming" and then
+            // pressing a movement key.
+            let arm = context
+                .spawn((Action::<ArmKick>::new(), arm_kick_binding))
+                .id();
+            context.spawn((
+                Action::<Kick>::new(),
+                Chord::single(arm),
+                Press::new(0.9),
+                kick_bindings,
+            ));
+        })),
     )
 }
 
-#[derive(Component, Clone, Copy, PartialEq, Eq, Hash)]
+fn apply_roll(roll: On<Fire<Roll>>, mut physics: Query<&mut PlayerPhysics>) {
+    let mut physics = physics.get_mut(roll.context).unwrap();
+    physics.velocity += roll.value;
+}
+
+fn apply_kick(kick: On<Fire<Kick>>, mut physics: Query<&mut PlayerPhysics>) {
+    let mut physics = physics.get_mut(kick.context).unwrap();
+    // Normalize the input to treat vectors that are barely inside the threshold
+    // the same way as a vector along the edge.
+    let dir = kick.value.normalize();
+
+    physics.velocity = dir * KICK_IMPULSE;
+}
+
+fn calculate_physics(time: Res<Time>, players: Query<(&mut Transform, &mut PlayerPhysics)>) {
+    for (mut transform, mut physics) in players {
+        // Apply velocity to transform
+        transform.translation += (physics.velocity * time.delta_secs()).extend(0.0);
+
+        // Apply friction to velocity
+        if physics.velocity.length_squared() > KINETIC_FRICTION_COEFF {
+            let friction_dir = physics.velocity.normalize() * -1.0;
+            physics.velocity += friction_dir * FRICTION * time.delta_secs();
+        }
+
+        // Check collision with walls and bounce
+        const BORDER_DIST: f32 = BORDER_WIDTH / 2.0 - BALL_RAD;
+        if transform.translation.x > BORDER_DIST {
+            transform.translation.x = BORDER_DIST;
+            physics.velocity.x *= -BOUNCINESS;
+        }
+        if transform.translation.x < -BORDER_DIST {
+            transform.translation.x = -BORDER_DIST;
+            physics.velocity.x *= -BOUNCINESS;
+        }
+        if transform.translation.y > BORDER_DIST {
+            transform.translation.y = BORDER_DIST;
+            physics.velocity.y *= -BOUNCINESS;
+        }
+        if transform.translation.y < -BORDER_DIST {
+            transform.translation.y = -BORDER_DIST;
+            physics.velocity.y *= -BOUNCINESS;
+        }
+    }
+}
+
+#[derive(Component)]
 enum Player {
     First,
     Second,
 }
 
-#[derive(Debug, InputAction)]
+#[derive(Component, Default)]
+struct PlayerPhysics {
+    velocity: Vec2,
+}
+
+#[derive(InputAction)]
 #[action_output(Vec2)]
-struct Movement;
+struct Roll;
+
+#[derive(InputAction)]
+#[action_output(Vec2)]
+struct Kick;
+
+#[derive(InputAction)]
+#[action_output(bool)]
+struct ArmKick;
